+++ build.config.ts
@@ -5,12 +5,13 @@ export default defineBuildConfig({
     { input: 'src/index', name: 'index' },
     { input: 'src/config', name: 'config' },
     { input: 'src/presets/minimal', name: 'presets/minimal' },
+    { input: 'src/api/index', name: 'api' },
     { input: 'src/cli', name: 'cli' },
   ],
   clean: true,
   declaration: true,
   rollup: {
-    emitCJS: false,
+    emitCJS: true,
     inlineDependencies: true,
   },
 })
