+++ package.json
@@ -2,7 +2,7 @@
   "name": "@vite-pwa/assets-generator",
   "type": "module",
   "version": "0.0.11",
-  "packageManager": "pnpm@8.10.5",
+  "packageManager": "pnpm@8.11.0",
   "description": "Zero-config PWA Assets Generator",
   "author": "antfu <anthonyfu117@hotmail.com>",
   "license": "MIT",
@@ -20,34 +20,68 @@
     "pwa icons generator",
     "pwa assets generator"
   ],
+  "sideEffects": false,
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs",
-      "default": "./dist/index.mjs"
+      "import": {
+        "types": "./dist/index.d.mts",
+        "default": "./dist/index.mjs"
+      },
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
+    },
+    "./api": {
+      "import": {
+        "types": "./dist/api.d.mts",
+        "default": "./dist/api.mjs"
+      },
+      "require": {
+        "types": "./dist/api.d.cts",
+        "default": "./dist/api.cjs"
+      }
     },
     "./config": {
-      "types": "./dist/config.d.ts",
-      "import": "./dist/config.mjs",
-      "default": "./dist/config.mjs"
+      "import": {
+        "types": "./dist/config.d.mts",
+        "default": "./dist/config.mjs"
+      },
+      "require": {
+        "types": "./dist/config.d.cts",
+        "default": "./dist/config.cjs"
+      }
     },
     "./presets/minimal": {
-      "types": "./dist/presets/minimal.d.ts",
-      "import": "./dist/presets/minimal.mjs",
-      "default": "./dist/presets/minimal.mjs"
-    }
+      "import": {
+        "types": "./dist/presets/minimal.d.mts",
+        "default": "./dist/presets/minimal.mjs"
+      },
+      "require": {
+        "types": "./dist/presets/minimal.d.cts",
+        "default": "./dist/presets/minimal.cjs"
+      }
+    },
+    "./*": "./*"
   },
-  "main": "dist/index.mjs",
+  "main": "dist/index.cjs",
   "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "*": [
+        "./dist/*",
+        "./*"
+      ]
+    }
+  },
   "bin": {
     "pwa-assets-generator": "./bin/pwa-assets-generator.mjs"
   },
   "files": [
     "bin",
     "dist",
-    "presets",
-    "*.d.ts"
+    "presets"
   ],
   "engines": {
     "node": ">=16.14.0"
@@ -64,18 +98,18 @@
     "cac": "^6.7.14",
     "colorette": "^2.0.20",
     "consola": "^3.2.3",
-    "sharp": "^0.32.6",
+    "sharp": "^0.33.0",
     "sharp-ico": "^0.1.5",
-    "unconfig": "^0.3.10"
+    "unconfig": "^0.3.11"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^0.43.1",
-    "@antfu/ni": "^0.21.9",
+    "@antfu/ni": "^0.21.12",
     "@types/debug": "^4.1.12",
     "@types/node": "^18.17.15",
     "bumpp": "^9.2.0",
     "eslint": "^8.53.0",
-    "typescript": "^5.2.2",
-    "unbuild": "^1.2.1"
+    "typescript": "^5.3.2",
+    "unbuild": "^2.0.0"
   }
 }
