+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './build'
+export * from './build.ts'
