+++ src/presets/index.ts
@@ -1,1 +1,1 @@
-export * from './minimal'
+export * from './minimal.ts'
